ultranova/requirements.txt

fastapi
uvicorn
gunicorn
pydantic
sqlalchemy
python-jose
psycopg2-binary
